chore(domain): move WatchOnlyError from watch-only to floresta-domain
The WatchOnlyError enum was previously defined inside floresta-watch-only.
This commit moves it to floresta-domain so it can be reused by other crates
without depending on the watch-only crate.
