+++ crates/floresta-node/src/florestad.rs
@@ -32,11 +32,11 @@ use floresta_common::try_and_log;
 use floresta_compact_filters::flat_filters_store::FlatFiltersStore;
 #[cfg(feature = "compact-filters")]
 use floresta_compact_filters::network_filters::NetworkFilters;
+use floresta_domain::wallet::error::WatchOnlyError;
 use floresta_electrum::electrum_protocol::ElectrumServer;
 use floresta_electrum::electrum_protocol::client_accept_loop;
 use floresta_mempool::Mempool;
 use floresta_watch_only::AddressCache;
-use floresta_watch_only::WatchOnlyError;
 use floresta_watch_only::kv_database::KvDatabase;
 use floresta_wire::UtreexoNodeConfig;
 use floresta_wire::address_man::AddressMan;
