+++ crates/floresta-node/src/json_rpc/res.rs
@@ -44,7 +44,7 @@ pub mod jsonrpc_interface {
     use floresta_chain::extensions::HeaderExtError;
     use floresta_common::impl_error_from;
     use floresta_domain::mempool::MempoolError;
-    use floresta_watch_only::WatchOnlyError;
+    use floresta_domain::wallet::error::WatchOnlyError;
     use floresta_wire::bitcoin_socket_addr::InvalidAddressError;
     use serde::Deserialize;
     use serde::Serialize;
